+++ database/create_table_activity_image_relation.sql
@@ -25,7 +25,7 @@ DROP TABLE IF EXISTS `activity_image_relation`;
 CREATE TABLE `activity_image_relation` (
   `activity_id` int(32) NOT NULL,
   `image_id` int(32) NOT NULL,
-  `generated_time` int(32) DEFAULT NULL,
+  `generated_time` bigint(20) DEFAULT NULL,
   KEY `activity_image_relation_ibfk_1` (`activity_id`),
   KEY `activity_image_relation_ibfk_2` (`image_id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
@@ -40,4 +40,4 @@ CREATE TABLE `activity_image_relation` (
 /*!40101 SET COLLATION_CONNECTION=@OLD_COLLATION_CONNECTION */;
 /*!40111 SET SQL_NOTES=@OLD_SQL_NOTES */;
 
--- Dump completed on 2014-11-11 12:37:14
+-- Dump completed on 2014-11-11 15:57:01
